+++ .eslintrc.js
@@ -1,0 +1,26 @@
+module.exports = {
+    root: true,
+    parserOptions: {
+      parser: 'babel-eslint'
+    },
+    env: {
+      browser: true,
+    },
+    extends: [
+      'plugin:vue/essential', 
+      'standard',
+      '@nextcloud'
+    ],
+    plugins: [
+      'vue'
+    ],
+    // add your custom rules here
+    rules: {
+      // allow async-await
+      'generator-star-spacing': 'off',
+      // allow debugger during development
+      'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+      'semi': [1, 'always']
+    }
+  };
+