+++ js/chunk-vendors.js
@@ -1,0 +1,1 @@
+../build/js/chunk-vendors.js