+++ src/main.js
@@ -1,0 +1,9 @@
+import Vue from 'vue';
+import App from './app.vue';
+import store from './store.js';
+
+export default new Vue({
+	store,
+	el: '#content',
+	render: h => h(App),
+});